// lib/database/models/lesson.model.ts
import { Document, Schema, model, models } from 'mongoose'

export interface ILesson extends Document {
	_id: string
	title: string
	description?: string
	location?: string
	createdAt: Date
	imageUrl: string
	startDateTime: Date
	endDateTime: Date
	price: string
	isFree: boolean
	url?: string
	category: { _id: string; name: string }
	organizer: { _id: string; firstName: string; lastName: string }
}

const LessonSchema = new Schema({
	title: { type: String, required: true },
	description: { type: String },
	location: { type: String },
	createdAt: { type: Date, default: Date.now },
	imageUrl: { type: String, required: true },
	startDateTime: { type: Date, default: Date.now },
	endDateTime: { type: Date, default: Date.now },
	price: { type: String },
	isFree: { type: Boolean, default: false },
	url: { type: String },
	category: { type: Schema.Types.ObjectId, ref: 'Category' },
	organizer: { type: Schema.Types.ObjectId, ref: 'User' },
})

const Lesson = models.Lesson || model('Lesson', LessonSchema)

export default Lesson